feat: Better navigation for looking up other profiles
